MatchMaker -- String Comparison for Variable Names
The library contains a set of functions implementing different levels of comparisons between variable names.
All the functions accept two arguments, name1 and name2, and return a similarity value in the range [0..1]. 0 means no relation, 1 means identity.
The operation of some functions may depend on various parameters. These have default values and can be set separately.
Global parameters:
- Case sensitivity. Default is not case sensitive.
- Word separators. Default is to take division of words into account, but not count _ as part of length of names.

For example, we want the comparison of  schoolBus  and  school_bus  to say they are the same. So _ matches lower-to-upper case change.
Internally can normalize all names to lowercase with _. This is a design decision, may depend on function.

The functions are:

edist -- edit distance.
Calculates the minimal number of add, delete, or change operations on single letters needed to change one name into the other.
Returns the quotient of the number of such operations divided by the length of the longer name.
Paramerter:
- Perform literal comparison, i.e. don't normalize names and count lower-upper and _ as changes.
Ideally would be good to also include move operation (equivalent to deleting a character in one place and adding it in another - can be implemented by post-processing the list of edits to find such pairs).

match -- wrapper of Python SequenceMatcher.
Returns the maximum of the default ratio function when applied in both directions.
Paramerter:
- Perform literal comparison, i.e. don't normalize names first.

all_match -- improved version of Python SequenceMatcher.
After each match is found, it is replaced by two different characters in the two strings and continue recursively.
This allows order change, i.e. matching the beginning of one name to the end of the other after matching the middle of both.
The added characters prevent matchings that are part here part there.
Need to make sure we compute ratio correctly in the end (that is, relative to the original length).
Parameters:
- Separator characters to use (should not appear in the names).
- Minimal length to go down to. Default 2. Going down to 1 will effectively find anagrams.

unedit_match -- same as above but without adding those characters.
This allows a sequence in one name to be partitioned across sides in the other name.
Rationale: like un-doing an edit that inserted the first match in the middle of the second.
Need to check if the result is indeed different, or actually it's the same thing.

words - service routine to partition names into words.
Knows camelCase and snake_case; handles all-uppercase acronyms and numbers.

words_match -- first partition the names into words, and then match words to each other including order differences.
The rational is that names like  kind_orange  and  kin_do_range  are different.
Parameters:
- Flag to treat numbers as separate words, ignore them, or just leave them.
- Degree of match for words. Default is to require perfect match (fast but may miss things). Option to allow partial match, but need not to go too low because that may become meaningless.
- Penalty if any for order changes.
- Take word length into account in computing match, or just number of words.
These are things you did in the last week, various things to discuss here.

semantic_match -- like above but also allow synonyms.
This is the new thing you started to add in the last few days.